Crate sea_orm_macros

Derive Macros

DeriveActiveEnum

A derive macro to implement sea_orm::ActiveEnum trait for enums.

DeriveActiveModel

The DeriveActiveModel derive macro will implement ActiveModelTrait for ActiveModel which provides setters and getters for all active values in the active model.

DeriveActiveModelBehavior

Models that a user can override

DeriveColumn

The DeriveColumn derive macro will implement [ColumnTrait] for Columns. It defines the identifier of each column by implementing Iden and IdenStatic. The EnumIter is also derived, allowing iteration over all enum variants.

DeriveCustomColumn

Derive a column if column names are not in snake-case

DeriveEntity

Create an Entity

DeriveEntityModel

This derive macro is the ‘almighty’ macro which automatically generates Entity, Column, and PrimaryKey from a given Model.

DeriveIntoActiveModel

Derive into an active model

DeriveMigrationName

The DeriveMigrationName derive macro will implement sea_orm_migration::MigrationName for a migration.

DeriveModel

The DeriveModel derive macro will implement ModelTrait for Model, which provides setters and getters for all attributes in the mod It also implements FromQueryResult to convert a query result into the corresponding Model.

DerivePrimaryKey

The DerivePrimaryKey derive macro will implement [PrimaryKeyToColumn] for PrimaryKey which defines tedious mappings between primary keys and columns. The [EnumIter] is also derived, allowing iteration over all enum variants.

DeriveRelation

The DeriveRelation derive macro will implement RelationTrait for Relation.

FromJsonQueryResult

FromQueryResult

Convert a query result into the corresponding Model.